Rework the filter handling on the search endpoints
Previously there were two quite similar ways of expressing filters:

  * simple_filter would take a Solr string and pass it straight to Solr
    as `fq`, but was hard to use if you were filtering by values with
    special characters (like quotes, parens, etc.)

  * filter_terms would take an array of JSON objects giving the fields
    and their values.  This was a bit weird, and only allowed you to
    construct filters that were effectively AND queries of terms.

We've merged these into a single "filter" parameter which takes a
JSONModel(:advanced_query).  This lets you use any type of query as a
filter, and we've enhanced the advanced_query_builder.rb file to make it
easy to build these.

On the frontend, filter_terms is still used in various places, but is
transformed into a "filter" parameter before sending to the backend.